// src/champion/dto/create-champion.dto.ts
export class CreateChampionDto {
  champNumber: number;
  name: string;
  worst1Name: string;
  worst2Name: string;
  worst3Name: string;
  worst1Rate: string;
  worst2Rate: string;
  worst3Rate: string;
  great1Name: string;
  great2Name: string;
  great3Name: string;
  great1Rate: string;
  great2Rate: string;
  great3Rate: string;
}